% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficas.R
\name{graficando_tabla}
\alias{graficando_tabla}
\title{Title Graficar tablas INEGI}
\usage{
graficando_tabla(bd, bd2, bd3, unidad_analisis)
}
\arguments{
\item{bd}{base con información de la primera variable INEGI seleccionada}

\item{bd2}{base con información de la segunda variable INEGI seleccionada}

\item{bd3}{base con información de la tercera variable INEGI seleccionada}

\item{unidad_analisis}{división cartográfica de interés}
}
\value{

}
\description{
Title Graficar tablas INEGI
}
