% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficas.R
\name{graficar_total_comparativo}
\alias{graficar_total_comparativo}
\title{Title}
\usage{
graficar_total_comparativo(info, analisis, mostrar = 7)
}
\arguments{
\item{analisis}{}
}
\value{

}
\description{
Title
}
