% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficas.R
\name{graficar_total_votacion}
\alias{graficar_total_votacion}
\title{Title}
\usage{
graficar_total_votacion(info)
}
\arguments{
\item{info}{base de datos procesada con los votos para cada partido y cada coalición, dividida en secciones, distritos y estado.}
}
\value{

}
\description{
Title
}
