% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informacion.R
\name{preparar_info_censo}
\alias{preparar_info_censo}
\title{Title Info censo}
\usage{
preparar_info_censo(
  inegi,
  unidad_analisis,
  id_unidad_analisis,
  variable,
  weight
)
}
\arguments{
\item{inegi}{Shapefile con información del censo por sección}

\item{unidad_analisis}{Nivel cartográfico de interés}
}
\value{

}
\description{
Title Info censo
}
