% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informacion.R
\name{preparar_info_de_eleccion}
\alias{preparar_info_de_eleccion}
\title{Title}
\usage{
preparar_info_de_eleccion(
  partidos,
  coaliciones,
  colores,
  bd,
  unidad_analisis,
  id_analisis,
  año_analisis,
  nombre_analisis
)
}
\arguments{
\item{bd}{}
}
\value{

}
\description{
Title
}
