% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informacion.R
\name{votos_coalicion}
\alias{votos_coalicion}
\title{Title}
\usage{
votos_coalicion(bd, partidos)
}
\arguments{
\item{bd}{Base de datos de votación.}

\item{partidos}{Vector de partidos que se encuentran coaligados. No importa el orden.}
}
\value{

}
\description{
Title
}
