% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficas.R
\name{graficar_fuerza_electoral}
\alias{graficar_fuerza_electoral}
\title{Title}
\usage{
graficar_fuerza_electoral(info, sf, analisis, interactiva = F)
}
\arguments{
\item{analisis}{}
}
\value{

}
\description{
Title
}
