% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informacion.R
\name{preparar_info_de_eleccion}
\alias{preparar_info_de_eleccion}
\title{La función básica del paquete. Ayuda a comunicar toda la información necesaria para los futuros procesamientos y gráficas.}
\usage{
preparar_info_de_eleccion(
  partidos,
  coaliciones,
  colores,
  bd,
  unidad_analisis,
  id_unidad_analisis,
  año_analisis,
  nombre_unidad_analisis
)
}
\arguments{
\item{partidos}{vector en minúscula de los partidos \strong{sin coalición} que participan en la elección.}

\item{coaliciones}{lista de vectores que se encuentran coaligados.}

\item{colores}{lista de vectores de colores con nombre. Los nombres corresponden al partido, o coalición para el que el color es asignado. Por mejorar.}

\item{bd}{base de datos con con formato homologado. Combinaciones de partidos separados por  '_'}
}
\value{

}
\description{
La función básica del paquete. Ayuda a comunicar toda la información necesaria para los futuros procesamientos y gráficas.
}
